feat: reward badge on track completed
