+++ apps/govquests-api/govquests/processes/lib/processes.rb
@@ -10,7 +10,7 @@ require_relative "processes/notify_on_reward_issued"
 require_relative "processes/notify_on_badge_earned"
 require_relative "processes/notify_on_tier_achieved"
 require_relative "processes/create_badge_on_track_or_quest_created"
-require_relative "processes/reward_badge_on_quest_completed"
+require_relative "processes/reward_badge_on_quest_or_track_completed"
 require_relative "processes/complete_track_on_quest_completed"
 
 require_relative "processes/deliver_notification_on_created"
@@ -22,7 +22,7 @@ module Processes
       UpdateProfileOnRewardIssued.new(event_store, command_bus).subscribe
       DistributeRewardsOnQuestCompleted.new(event_store, command_bus).subscribe
       CreateBadgeOnTrackOrQuestCreated.new(event_store, command_bus).subscribe
-      RewardBadgeOnQuestCompleted.new(event_store, command_bus).subscribe
+      RewardBadgeOnQuestOrTrackCompleted.new(event_store, command_bus).subscribe
       CompleteTrackOnQuestCompleted.new(event_store, command_bus).subscribe
 
       NotifyOnQuestCompleted.new(event_store, command_bus).subscribe
